+++ config.toml
@@ -1,0 +1,19 @@
+foo="bar"
+
+[Main]
+
+application_mode = "dev"
+log_level = "debug"
+log_destination = "stdout"
+log_in_json = 0
+log_sql = 0
+timezone = +3
+enable_swagger = 1
+swagger_doc_url = "/doc"
+swagger_redoc_url = "/redoc"
+enable_security = 1
+
+[AdminGUI]
+admin_enable = 1
+admin_url = "/admin/"
+
