+++ configuration/__init__.py
@@ -1,4 +1,4 @@
-from .loader import Configuration
+from .model import Configuration
 from os import environ
 config = Configuration()
 config.load(environ['X_FA_CONF_FILE'])
